Initial commit
Sample that demonstrates how both Kafka and Kafka Streams binders
can co-exist in the same app. It also shows the right usage of including
the new test binder.

Also added notes as comments for using the old test binder (spring-cloud-stream-test-support).
See the maven pom.xml where this dependency is commented out and the comments in the tests
on how to exclude the auto configuration for the test binder.